refactor : 경로 조회 SectionEdge 종속성 분리
