+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -4,6 +4,7 @@ import nextstep.subway.line.domain.Line;
 import nextstep.subway.station.domain.Station;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static nextstep.subway.path.application.exception.InvalidPathException.*;
 
@@ -31,5 +32,22 @@ public class PathFinder implements PathFinderStrategy {
         if (graph.isNotConnectable(source, target)) {
             throw NOT_CONNECTABLE;
         }
+    }
+
+    @Override
+    public List<Line> lineOnPath(List<Line> lines) {
+        return graph.getEdges().stream()
+                .map(edge -> findBy(lines, edge))
+                .collect(Collectors.toList());
+    }
+
+    private Line findBy(List<Line> lines, SectionEdge edge) {
+        Station source = (Station) edge.getSource();
+        Station target = (Station) edge.getTarget();
+
+        return lines.stream()
+                .filter(line -> line.matchSection(source, target))
+                .findFirst()
+                .orElseThrow(() -> NOT_EXIST_STATION);
     }
 }
