+++ src/main/java/nextstep/subway/path/domain/PathFinderStrategy.java
@@ -1,9 +1,14 @@
 package nextstep.subway.path.domain;
 
+import nextstep.subway.line.domain.Line;
 import nextstep.subway.station.domain.Station;
+
+import java.util.List;
 
 public interface PathFinderStrategy {
 
     Path shortestPath(Station source, Station target);
+
+    List<Line> lineOnPath(List<Line>lines);
 
 }
